doit_news/requirements.txt
asgiref==3.7.2
Django==5.0.3
factory-boy==3.3.0
Faker==24.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.16.0
sqlparse==0.4.4
tzdata==2024.1
